Consolidate Orders list header variants
